chore: adding security to contract
