+++ contract/package.json
