+++ contract/programs/token_contract/Cargo.toml
@@ -18,3 +18,5 @@ default = []
 [dependencies]
 anchor-lang = "0.29.0"
 anchor-spl = "0.29.0"
+solana-program = "=1.18.2"
+solana-security-txt = "1.1.1"
