+++ contract/programs/token_contract/src/instructions/mint.rs
@@ -1,0 +1,36 @@
+use anchor_lang::prelude::*;
+use anchor_spl::token;
+use anchor_spl::token::{Token, MintTo};
+
+#[derive(Accounts)]
+pub struct MintToken<'info> {
+    /// CHECK: This is the token that we want to mint
+    #[account(mut)]
+    pub mint: UncheckedAccount<'info>,
+    pub token_program: Program<'info, Token>,
+    /// CHECK: This is the token account that we want to mint tokens to
+    #[account(mut)]
+    pub token_account: UncheckedAccount<'info>,
+    /// CHECK: the authority of the mint account
+    #[account(mut)]
+    pub authority: AccountInfo<'info>,
+}
+
+
+pub fn mint_token(ctx: Context<MintToken>, amount: u64) -> Result<()> {
+    // Create the MintTo struct for our context
+    let cpi_accounts = MintTo {
+        mint: ctx.accounts.mint.to_account_info(),
+        to: ctx.accounts.token_account.to_account_info(),
+        authority: ctx.accounts.authority.to_account_info(),
+    };
+    
+    let cpi_program = ctx.accounts.token_program.to_account_info();
+    // Create the CpiContext we need for the request
+    let cpi_ctx = CpiContext::new(cpi_program, cpi_accounts);
+
+    // Execute anchor's helper function to mint tokens
+    token::mint_to(cpi_ctx, amount)?;
+    
+    Ok(())
+}
