+++ contract/programs/token_contract/src/instructions/mod.rs
@@ -1,0 +1,5 @@
+pub mod transfer;
+pub mod mint;
+
+pub use mint::*;
+pub use transfer::*;
