+++ contract/programs/token_contract/src/instructions/transfer.rs
@@ -1,0 +1,34 @@
+use anchor_lang::prelude::*;
+use anchor_spl::token::{Token, Transfer};
+
+#[derive(Accounts)]
+pub struct TransferToken<'info> {
+    pub token_program: Program<'info, Token>,
+    /// CHECK: The associated token account that we are transferring the token from
+    #[account(mut)]
+    pub from: UncheckedAccount<'info>,
+    /// CHECK: The associated token account that we are transferring the token to
+    #[account(mut)]
+    pub to: AccountInfo<'info>,
+    // the authority of the from account 
+    pub from_authority: Signer<'info>,
+}
+
+
+pub fn transfer_token(ctx: Context<TransferToken>, amount: u64) -> Result<()> {
+    // Create the Transfer struct for our context
+    let transfer_instruction = Transfer{
+        from: ctx.accounts.from.to_account_info(),
+        to: ctx.accounts.to.to_account_info(),
+        authority: ctx.accounts.from_authority.to_account_info(),
+    };
+     
+    let cpi_program = ctx.accounts.token_program.to_account_info();
+    // Create the Context for our Transfer request
+    let cpi_ctx = CpiContext::new(cpi_program, transfer_instruction);
+
+    // Execute anchor's helper function to transfer tokens
+    anchor_spl::token::transfer(cpi_ctx, amount)?;
+
+    Ok(())
+}
