+++ package.json
@@ -35,7 +35,7 @@
     "apps/*",
     "packages/*",
     "backend/*",
-    "contract/*",
+    "contract",
     "openapi",
     "monitor"
   ],
